requirements.txt
Pillow>=10.0.0
openpyxl>=3.1.0
reportlab>=4.0.0
pyinstaller>=6.0.0
